feat: leverage version generation as a library
This gets more accurate info since a crate will regenerate version info as part of its build process.
